% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkolmo.R
\name{dkolmo}
\alias{dkolmo}
\title{Kolmogorov Distribution: Probability Density Function}
\usage{
dkolmo(x, log = FALSE)
}
\arguments{
\item{x}{The argument whose PDF is to be computed}

\item{log}{If it is true then log-density is returned}
}
\value{
Values of Kolmogorov PDF
}
\description{
Evaluation of the Kolmogorov probability density function (PDF).
}
\details{
The approach described in Onorati and Liseo (2022) is used. There are two
expressions for the Kolmogorov density. One representation is used  if
0 < x < 1.207 and the other for x >= 1.207. In both cases the series is
truncated up to the first 15 terms.
}
\references{
Onorati, P., & Liseo, B. (2022). A Random Number Generator for
the Kolmogorov Distribution. arXiv preprint arXiv:2208.13598.
}
