% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompNegbinom_rng.R
\name{GompNegbinom_rng}
\alias{GompNegbinom_rng}
\title{Gompertz Model with Negative Binomial Sampling Error: Random Number
Generator}
\usage{
GompNegbinom_rng = function(T, theta1, theta2, b, prob)
}
\arguments{
\item{T}{The length of the chain, say the sample size.}

\item{theta1}{The mean of the stationary distribution.}

\item{theta2}{The variance of the stationary distribution.}

\item{b}{The autoregressive coefficient.}

\item{prob}{Probability parameter for negative binomial.}
}
\value{
A vector containing the noised chain.
}
\description{
Generation of independent draws from Gompertz model with negative binomial
sampling error distribution.
}
\details{
This function sample from the following model: \deqn{ N^\star_t |
N_t \sim \mathrm{negbinom}(p = prob, mu = N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, . }

A different parametrization is used, say: \deqn{ \theta_1 = -\frac{a}{b} \, ,
 \, \theta_2 = -\frac{\sigma^2}{b(2 + b)} \, , } which implies \deqn{ Z_1,
 Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, , \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, . }

As \code{prob} goes to one, the model converges to the Gompertz Poisson
sampling error distribution
}
\references{
\link{INSERT REFERENCE}
}
