% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompPois_likelihood.R
\name{GompPois_likelihood}
\alias{GompPois_likelihood}
\title{Gompertz Model with Poisson Sampling Error: Likelihood Computation}
\usage{
GompPois_likelihood = function(Nstar, theta1, theta2, b,
                               nsim = 1e+5, log = FALSE, details = FALSE)
}
\arguments{
\item{Nstar}{The vector of the data.}

\item{theta1}{The mean of stationary distribution whose likelihood must be
computed.}

\item{theta2}{The variance of stationary distribution whose likelihood must
be computed.}

\item{b}{The autoregressive coefficient whose likelihood must be computed.}

\item{nsim}{The number of simulations for the importance sampling.}

\item{log}{If it is true then log-density is returned.}

\item{details}{If it is true then also the weights and draws are returned.}
}
\value{
Likelihood of the arguments and, optionally, weights and draws from
importance sampling.
}
\description{
Evaluation of the likelihood for Gompertz model with Poisson
sampling error distribution
}
\details{
This function computes the likelihood for the following model:
\deqn{
 N^\star_t | N_t \sim \mathrm{Poisson}(N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, .
 }

A different parametrization is used, say:
\deqn{
 \theta_1 = -\frac{a}{b} \, , \, \theta_2 = -\frac{\sigma^2}{b(2 + b)}
 \, ,
 }
which implies
\deqn{
 Z_1, Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, , \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, .
}
Thus, the likelihood function is \eqn{\pi(N^\star | \theta_1, \theta_2, b)}.
The loglikelihood is computed through an importance sampling algorithm.
The importance density is constructed via a sequence of Gaussian
distributions minimizing the Kolmogorov-Smirnov distance.

Notice that a simple close formula exists for the parameters of the Gaussian
distributions. See \link{INSERT REFERENCE} for more details.
}
\references{
\link{INSERT REFERENCE}
}
