% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompPois_mcmcEM.R
\name{GompPois_mcmcEM}
\alias{GompPois_mcmcEM}
\title{Gompertz Model with Poisson Sampling Error: Markov Chain Monte Carlo
Expectation Maximization Algorithm}
\usage{
GompPois_mcmcEM = function(Nstar, maxit = 1e+3,
                            abstol = sqrt(.Machine$double.eps),
                            max_nsim = 1e+5, nsim = 1e+4,
                            starter = NULL, verbose = +Inf)
}
\arguments{
\item{Nstar}{The vector of the data.}

\item{maxit}{The maximum number of iterations.}

\item{abstol}{The absolute convergence tolerance.}

\item{max_nsim}{The maximum number of complete scans for MCMC step.}

\item{nsim}{The initial number of complete scans for MCMC step.}

\item{starter}{The list providing the starter point.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
MLE estimates.
}
\description{
Computation of the MLE for Gompertz model with Poisson sampling error
distribution
}
\details{
This function computes the MLE for the following model:
\deqn{
 N^\star_t | N_t \sim \mathrm{Poisson}(N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, .
 }

A different parametrization is used, say:
\deqn{
 \theta_1 = -\frac{a}{b} \, , \, \theta_2 = -\frac{\sigma^2}{b(2 + b)}
 \, ,
 }
which implies
\deqn{
 Z_1, Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, .
}
The following scheme is used
\itemize{
\item for \eqn{t = 1, 2, ..., T}:
sample \eqn{Z_t \vert N^\star_t, Z_{-t}, b, \theta_1, \theta_2}
(using acceptance-rejection) and ...

\item maximize the log-likelihood by taking ...
}

\code{starter} must be a list with the following structure:
\code{starter$theta1}, \code{starter$theta2}, \code{starter$b}, or
\code{NULL}. In the latter case the starting point is generated by method of
moments.
}
\references{
\link{INSERT REFERENCE}
}
