% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompPois_MoM.R
\name{GompPois_MoM}
\alias{GompPois_MoM}
\title{Gompertz Model with Poisson Sampling Error: Method of Moments}
\usage{
GompPois_MoM = function(Nstar)
}
\arguments{
\item{Nstar}{The vector of the data.}
}
\value{
Named list of estimates.
}
\description{
Point estimator using method of moments for Gompertz model with Poisson
sampling error distribution.
}
\details{
This function estimates the parameters of the following model:
\deqn{
 N^\star_t | N_t \sim \mathrm{Poisson}(N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, .
 }

A different parametrization is used, say:
\deqn{
 \theta_1 = -\frac{a}{b} \, , \, \theta_2 = -\frac{\sigma^2}{b(2 + b)}
 \, ,
 }
which implies
\deqn{
 Z_1, Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, , \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, .
}
The method of moments is used. In particular, the parameters are estimated
such that the mean, variance, and autocovariance at lag 1 are matched to the
empirical one.
}
\references{
\link{INSERT REFERENCE}
}
