% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompPois_stEM_eb.R
\name{GompPois_stEM_eb}
\alias{GompPois_stEM_eb}
\title{Gompertz Model with Poisson Sampling Error with Stochastic
Expectation Maximization: Empirical Bayes}
\usage{
GompPois_stEM_eb = function(nsim, Nstar, starter = NULL, burn = 0,
                            thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations for the Markov chain.}

\item{Nstar}{The vector of the data.}

\item{phi1}{The fixed value of shape parameter of inverse gamma prior.}

\item{eta1}{The fixed value of location parameter of normal prior.}

\item{starter}{List for starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}

\item{includeMCMC}{If it is true then MCMC output is included.}
}
\value{
Return a list containing the mean of the sequence of optimizers for
scale parameters and fixed values of shape and location parameters.
}
\description{
Calibration of hyper-parameters of the NIG prior for Gompertz model with
Poisson sampling error distribution using stochastic EM algorithm.
}
\details{
This function calibrates the hyper-parameters of the NIG prior for the
following model:
\deqn{
 N^\star_t | N_t \sim \mathrm{Poisson}(N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, .
 }

A different parametrization is used, say:
\deqn{
 \theta_1 = -\frac{a}{b} \, , \, \theta_2 = -\frac{\sigma^2}{b(2 + b)}
 \, ,
 }
which implies
\deqn{
 Z_1, Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, , \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, .
}
The priors are:
\deqn{
 b \sim U(-2, 0) \, , \\
 \theta_2 \sim Inv.Gamma(\phi_1, \phi_2) \, , \\
 \theta_1 \vert \theta_2 \sim N(\eta_1, \eta_2 \theta_2) \, .
}
Setting \eqn{\beta = \log(-b / (1 + b))}, it is obtained \eqn{\beta \sim
Logis(0, 1)}, that is \eqn{\beta \vert V \sim N(0, V)} and \eqn{V} is a
logistic Kolmogorov distribution (four times the square of a Kolmogorov
distribution).
The values of \eqn{\phi_1} and \eqn{\eta_1} are kept fixed while a stochastic
expectation maximization algorithm is used to calibrate \eqn{\phi_2, \eta_2}.
}
\references{
\link{INSERT REFERENCE}
}
