% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GompPois_stEM_mle.R
\name{GompPois_stEM_mle}
\alias{GompPois_stEM_mle}
\title{Gompertz Model with Poisson Sampling Error with Stochastic
Expectation Maximization: Maximum Likelihood Estimator}
\usage{
GompPois_stEM_mle = function(nsim, Nstar, starter = NULL, burn = 1e+3,
                             thin = 1, verbose = +Inf)
}
\arguments{
\item{nsim}{The number of simulations for the Markov chain.}

\item{Nstar}{The vector of the data.}

\item{starter}{List for starting point.}

\item{burn}{The number of draws to be discarded as burn-in.}

\item{thin}{The thinning parameter.}

\item{verbose}{The period for printing status of the chain.}
}
\value{
Return a matrix containing a sequence of optimizers of the
parameters.
}
\description{
Computation of the maximum likelihood estimator for Gompertz model with
Poisson sampling error distribution using stochastic EM algorithm.
}
\details{
This function computes the maximum likelihood estimator for the following
model:
\deqn{
 N^\star_t | N_t \sim \mathrm{Poisson}(N_t) \, , \\
 N_t = \exp(Z_t) \, , \\
 Z_t = a + (1 + b) Z_{t - 1} + \varepsilon_t \, , \\
 \varepsilon_t \overset{i.i.d.}{\sim} N(0, \sigma^2) \, .
 }

A different parametrization is used, say:
\deqn{
 \theta_1 = -\frac{a}{b} \, , \, \theta_2 = -\frac{\sigma^2}{b(2 + b)}
 \, ,
 }
which implies
\deqn{
 Z_1, Z_2, ..., Z_T \sim N_T ( \theta_1 1_T, \theta_2 \bar{\Sigma})  \, , \\
 \bar{\Sigma}_{ij} = (1 + b ) ^ {\vert i - j \vert} \, .
}
Thus, a stochastic expectation maximization algorithm is used.
See \link{INSERT REFERENCE} for more details.
}
\references{
\link{INSERT REFERENCE}
}
