% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpostlogiskolmo.R
\name{rpostlogiskolmo}
\alias{rpostlogiskolmo}
\title{Posterior Logistic Kolmogorov Distribution: Random Number Generator}
\usage{
rpostlogiskolmo = function(n, x)
}
\arguments{
\item{n}{The number of simulations.}

\item{x}{Vector of conditioning values.}
}
\value{
Vector of draws from posterior logistic Kolmogorov distribution.
}
\description{
Exact random number generator for posterior logistic Kolmogorov distribution.
}
\details{
It is said \eqn{W \sim \mathrm{LK}(\cdot)}, say a logistic Kolmogorov
distribution, if \eqn{W = 4 V^2} and \eqn{V} follows a Kolmogorov
distribution.

The posterior logistic Kolmogorov distribution is the random variable
\eqn{W \vert X} where \eqn{W \sim \mathrm{LK}(\cdot)} and
\eqn{X \vert W \sim N(0, W)}.

It is used the acceptance-rejection scheme described in Onorati and Liseo
(2022)
}
\references{
Onorati, P., & Liseo, B. (2022). An extension of the unified skew-normal
family of distributions and application to Bayesian binary regression.
arXiv preprint arXiv:2209.03474.
}
